Move pytest fixtures to separate file
