+++ tests/test_database.py
@@ -1,30 +1,11 @@
-import os
 from datetime import datetime
 import pytz
-
-import pytest
 from flask_sqlalchemy import SQLAlchemy
 
-from statok_app.models.database import db
 from statok_app.models.category import Category, CategoryType
 from statok_app.models.operation import Operation, get_current_time
-from statok_app.app import create_test_app
 
-
-MYSQL_USER = os.environ.get("MYSQL_USER")
-MYSQL_PASS = os.environ.get("MYSQL_PASS")
-MYSQL_DBNAME = "statok_test"
-
-@pytest.fixture
-def test_db() -> SQLAlchemy:
-    test_app = create_test_app()
-    test_app.config["TESTING"] = True
-
-    with test_app.app_context():
-        db.create_all()
-        yield db
-        db.session.remove()
-        db.drop_all()
+from test_fixtures import test_db, dummy_db
 
 
 def test_db_insert_category(test_db: SQLAlchemy):
